+++ source/common/config/grpc_mux_impl.h
@@ -36,6 +36,7 @@ public:
                             GrpcMuxCallbacks& callbacks) override;
   void pause(const std::string& type_url) override;
   void resume(const std::string& type_url) override;
+  bool paused(const std::string& type_url) const override;
 
   void sendDiscoveryRequest(const std::string& type_url);
 
@@ -120,8 +121,15 @@ public:
                             GrpcMuxCallbacks&) override {
     throw EnvoyException("ADS must be configured to support an ADS config source");
   }
-  void pause(const std::string&) override {}
-  void resume(const std::string&) override {}
+  void pause(const std::string& type_url) override { paused_apis_.insert(type_url); }
+  void resume(const std::string& type_url) override { paused_apis_.erase(type_url); }
+  bool paused(const std::string& type_url) const override {
+    return paused_apis_.find(type_url) != paused_apis_.end();
+  }
+
+private:
+  // Helpful for testing ClusterManagerImpl.
+  std::unordered_set<std::string> paused_apis_;
 };
 
 } // namespace Config