+++ test/common/upstream/cds_api_impl_test.cc
@@ -6,13 +6,10 @@
 #include "envoy/api/v2/core/config_source.pb.validate.h"
 
 #include "common/config/utility.h"
-#include "common/http/message_impl.h"
-#include "common/json/json_loader.h"
 #include "common/protobuf/utility.h"
 #include "common/upstream/cds_api_impl.h"
 
 #include "test/common/upstream/utility.h"
-#include "test/mocks/local_info/mocks.h"
 #include "test/mocks/protobuf/mocks.h"
 #include "test/mocks/upstream/mocks.h"
 #include "test/test_common/printers.h"
@@ -38,34 +35,14 @@ MATCHER_P(WithName, expectedName, "") { return arg.name() == expectedName; }
 
 class CdsApiImplTest : public testing::Test {
 protected:
-  CdsApiImplTest()
-      : request_(&cm_.async_client_), api_(Api::createApiForTest(store_)),
-        cds_version_(
-            store_.gauge("cluster_manager.cds.version", Stats::Gauge::ImportMode::NeverImport)) {}
-
   void setup() {
-    const std::string config_yaml = R"EOF(
-api_config_source:
-  cluster_names:
-  - foo_cluster
-  refresh_delay: 1s
-  api_type: REST
-    )EOF";
-
     envoy::api::v2::core::ConfigSource cds_config;
-    TestUtility::loadFromYamlAndValidate(config_yaml, cds_config);
-    cluster_map_.emplace("foo_cluster", mock_cluster_);
-    EXPECT_CALL(cm_, clusters()).WillRepeatedly(Return(cluster_map_));
-    EXPECT_CALL(mock_cluster_, info()).Times(AnyNumber());
-    EXPECT_CALL(*mock_cluster_.info_, addedViaApi());
-    EXPECT_CALL(mock_cluster_, info()).Times(AnyNumber());
-    EXPECT_CALL(*mock_cluster_.info_, type());
-    cds_ = CdsApiImpl::create(cds_config, cm_, dispatcher_, random_, local_info_, store_,
-                              validation_visitor_, *api_);
+    cds_ = CdsApiImpl::create(cds_config, cm_, store_, validation_visitor_);
     cds_->setInitializedCb([this]() -> void { initialized_.ready(); });
 
-    expectRequest();
+    EXPECT_CALL(*cm_.subscription_factory_.subscription_, start(_));
     cds_->initialize();
+    cds_callbacks_ = cm_.subscription_factory_.callbacks_;
   }
 
   void expectAdd(const std::string& cluster_name, const std::string& version = std::string("")) {
@@ -77,58 +54,22 @@ api_config_source:
         .WillOnce(Throw(EnvoyException(exception_msg)));
   }
 
-  void expectRequest() {
-    EXPECT_CALL(cm_, httpAsyncClientForCluster("foo_cluster"));
-    EXPECT_CALL(cm_.async_client_, send_(_, _, _))
-        .WillOnce(
-            Invoke([&](Http::MessagePtr& request, Http::AsyncClient::Callbacks& callbacks,
-                       const Http::AsyncClient::RequestOptions&) -> Http::AsyncClient::Request* {
-              EXPECT_EQ(
-                  (Http::TestHeaderMapImpl{
-                      {":method", "POST"},
-                      {":path", "/v2/discovery:clusters"},
-                      {":authority", "foo_cluster"},
-                      {"content-type", "application/json"},
-                      {"content-length",
-                       request->body() ? fmt::format_int(request->body()->length()).str() : "0"}}),
-                  request->headers());
-              callbacks_ = &callbacks;
-              return &request_;
-            }));
-  }
-
-  ClusterManager::ClusterInfoMap makeClusterMap(std::vector<std::string> clusters) {
+  ClusterManager::ClusterInfoMap makeClusterMap(const std::vector<std::string>& clusters) {
     ClusterManager::ClusterInfoMap map;
-    for (auto cluster : clusters) {
+    for (const auto& cluster : clusters) {
       map.emplace(cluster, cm_.thread_local_cluster_.cluster_);
     }
     return map;
   }
 
-  static Http::MessagePtr parseResponseMessageFromYaml(const std::string& yaml) {
-    Http::MessagePtr message(new Http::ResponseMessageImpl(
-        Http::HeaderMapPtr{new Http::TestHeaderMapImpl{{":status", "200"}}}));
-    const auto json = Json::Factory::loadFromYamlString(yaml);
-    ASSERT(json->isObject());
-    message->body() = std::make_unique<Buffer::OwnedImpl>(json->asJsonString());
-    return message;
-  }
-
   NiceMock<MockClusterManager> cm_;
   Upstream::ClusterManager::ClusterInfoMap cluster_map_;
   Upstream::MockClusterMockPrioritySet mock_cluster_;
-  NiceMock<Event::MockDispatcher> dispatcher_;
-  NiceMock<Runtime::MockRandomGenerator> random_;
-  NiceMock<LocalInfo::MockLocalInfo> local_info_;
   Stats::IsolatedStoreImpl store_;
-  Http::MockAsyncClientRequest request_;
   CdsApiPtr cds_;
-  Event::MockTimer* interval_timer_;
-  Http::AsyncClient::Callbacks* callbacks_{};
+  Config::SubscriptionCallbacks* cds_callbacks_{};
   ReadyWatcher initialized_;
   NiceMock<ProtobufMessage::MockValidationVisitor> validation_visitor_;
-  Api::ApiPtr api_;
-  Stats::Gauge& cds_version_;
 };
 
 // Negative test for protoc-gen-validate constraints.
@@ -143,14 +84,12 @@ TEST_F(CdsApiImplTest, ValidateFail) {
 
   EXPECT_CALL(cm_, clusters()).WillRepeatedly(Return(cluster_map_));
   EXPECT_CALL(initialized_, ready());
-  EXPECT_THROW(dynamic_cast<CdsApiImpl*>(cds_.get())->onConfigUpdate(clusters, ""), EnvoyException);
-  EXPECT_CALL(request_, cancel());
+  EXPECT_THROW(cds_callbacks_->onConfigUpdate(clusters, ""), EnvoyException);
 }
 
 // Regression test against only updating versionInfo() if at least one cluster
 // is are added/updated even if one or more are removed.
 TEST_F(CdsApiImplTest, UpdateVersionOnClusterRemove) {
-  interval_timer_ = new Event::MockTimer(&dispatcher_);
   InSequence s;
 
   setup();
@@ -165,34 +104,24 @@ resources:
     eds_config:
       path: eds path
 )EOF";
+  auto response1 = TestUtility::parseYaml<envoy::api::v2::DiscoveryResponse>(response1_yaml);
 
   EXPECT_CALL(cm_, clusters()).WillOnce(Return(ClusterManager::ClusterInfoMap{}));
   expectAdd("cluster1", "0");
   EXPECT_CALL(initialized_, ready());
-  EXPECT_CALL(*interval_timer_, enableTimer(_));
   EXPECT_EQ("", cds_->versionInfo());
-  EXPECT_EQ(
-      0UL,
-      store_.gauge("cluster_manager.cds.version", Stats::Gauge::ImportMode::NeverImport).value());
 
-  callbacks_->onSuccess(parseResponseMessageFromYaml(response1_yaml));
+  cds_callbacks_->onConfigUpdate(response1.resources(), response1.version_info());
   EXPECT_EQ("0", cds_->versionInfo());
-
-  expectRequest();
-  interval_timer_->callback_();
 
   const std::string response2_yaml = R"EOF(
 version_info: '1'
 resources:
 )EOF";
+  auto response2 = TestUtility::parseYaml<envoy::api::v2::DiscoveryResponse>(response2_yaml);
   EXPECT_CALL(cm_, clusters()).WillOnce(Return(makeClusterMap({"cluster1"})));
-
   EXPECT_CALL(cm_, removeCluster("cluster1")).WillOnce(Return(true));
-  EXPECT_CALL(*interval_timer_, enableTimer(_));
-  callbacks_->onSuccess(parseResponseMessageFromYaml(response2_yaml));
-
-  EXPECT_EQ(2UL, store_.counter("cluster_manager.cds.update_attempt").value());
-  EXPECT_EQ(2UL, store_.counter("cluster_manager.cds.update_success").value());
+  cds_callbacks_->onConfigUpdate(response2.resources(), response2.version_info());
   EXPECT_EQ("1", cds_->versionInfo());
 }
 
@@ -210,11 +139,9 @@ TEST_F(CdsApiImplTest, ValidateDuplicateClusters) {
 
   EXPECT_CALL(cm_, clusters()).WillRepeatedly(Return(cluster_map_));
   EXPECT_CALL(initialized_, ready());
-  EXPECT_THROW_WITH_MESSAGE(dynamic_cast<CdsApiImpl*>(cds_.get())->onConfigUpdate(clusters, ""),
-                            EnvoyException,
+  EXPECT_THROW_WITH_MESSAGE(cds_callbacks_->onConfigUpdate(clusters, ""), EnvoyException,
                             "Error adding/updating cluster(s) duplicate_cluster: duplicate cluster "
                             "duplicate_cluster found");
-  EXPECT_CALL(request_, cancel());
 }
 
 TEST_F(CdsApiImplTest, EmptyConfigUpdate) {
@@ -224,10 +151,9 @@ TEST_F(CdsApiImplTest, EmptyConfigUpdate) {
 
   EXPECT_CALL(cm_, clusters()).WillOnce(Return(ClusterManager::ClusterInfoMap{}));
   EXPECT_CALL(initialized_, ready());
-  EXPECT_CALL(request_, cancel());
 
   Protobuf::RepeatedPtrField<ProtobufWkt::Any> clusters;
-  dynamic_cast<CdsApiImpl*>(cds_.get())->onConfigUpdate(clusters, "");
+  cds_callbacks_->onConfigUpdate(clusters, "");
 }
 
 TEST_F(CdsApiImplTest, ConfigUpdateWith2ValidClusters) {
@@ -238,7 +164,6 @@ TEST_F(CdsApiImplTest, ConfigUpdateWith2ValidClusters) {
 
   EXPECT_CALL(cm_, clusters()).WillOnce(Return(ClusterManager::ClusterInfoMap{}));
   EXPECT_CALL(initialized_, ready());
-  EXPECT_CALL(request_, cancel());
 
   Protobuf::RepeatedPtrField<ProtobufWkt::Any> clusters;
 
@@ -252,7 +177,7 @@ TEST_F(CdsApiImplTest, ConfigUpdateWith2ValidClusters) {
   clusters.Add()->PackFrom(cluster_2);
   expectAdd("cluster_2");
 
-  dynamic_cast<CdsApiImpl*>(cds_.get())->onConfigUpdate(clusters, "");
+  cds_callbacks_->onConfigUpdate(clusters, "");
 }
 
 TEST_F(CdsApiImplTest, DeltaConfigUpdate) {
@@ -261,7 +186,6 @@ TEST_F(CdsApiImplTest, DeltaConfigUpdate) {
     setup();
   }
   EXPECT_CALL(initialized_, ready());
-  EXPECT_CALL(request_, cancel());
 
   {
     Protobuf::RepeatedPtrField<envoy::api::v2::Resource> resources;
@@ -283,7 +207,7 @@ TEST_F(CdsApiImplTest, DeltaConfigUpdate) {
       resource->set_name("cluster_2");
       resource->set_version("v1");
     }
-    dynamic_cast<CdsApiImpl*>(cds_.get())->onConfigUpdate(resources, {}, "v1");
+    cds_callbacks_->onConfigUpdate(resources, {}, "v1");
   }
 
   {
@@ -300,7 +224,7 @@ TEST_F(CdsApiImplTest, DeltaConfigUpdate) {
     Protobuf::RepeatedPtrField<std::string> removed;
     *removed.Add() = "cluster_1";
     EXPECT_CALL(cm_, removeCluster(StrEq("cluster_1"))).WillOnce(Return(true));
-    dynamic_cast<CdsApiImpl*>(cds_.get())->onConfigUpdate(resources, removed, "v2");
+    cds_callbacks_->onConfigUpdate(resources, removed, "v2");
   }
 }
 
@@ -312,7 +236,6 @@ TEST_F(CdsApiImplTest, ConfigUpdateAddsSecondClusterEvenIfFirstThrows) {
 
   EXPECT_CALL(cm_, clusters()).WillOnce(Return(ClusterManager::ClusterInfoMap{}));
   EXPECT_CALL(initialized_, ready());
-  EXPECT_CALL(request_, cancel());
 
   Protobuf::RepeatedPtrField<ProtobufWkt::Any> clusters;
 
@@ -332,29 +255,11 @@ TEST_F(CdsApiImplTest, ConfigUpdateAddsSecondClusterEvenIfFirstThrows) {
   expectAddToThrow("cluster_3", "Another exception");
 
   EXPECT_THROW_WITH_MESSAGE(
-      dynamic_cast<CdsApiImpl*>(cds_.get())->onConfigUpdate(clusters, ""), EnvoyException,
+      cds_callbacks_->onConfigUpdate(clusters, ""), EnvoyException,
       "Error adding/updating cluster(s) cluster_1: An exception, cluster_3: Another exception");
 }
 
-TEST_F(CdsApiImplTest, InvalidOptions) {
-  const std::string config_yaml = R"EOF(
-api_config_source:
-  cluster_names:
-  - foo_cluster
-  refresh_delay: 1s
-  )EOF";
-
-  local_info_.node_.set_cluster("");
-  local_info_.node_.set_id("");
-  envoy::api::v2::core::ConfigSource cds_config;
-  TestUtility::loadFromYamlAndValidate(config_yaml, cds_config);
-  EXPECT_THROW(CdsApiImpl::create(cds_config, cm_, dispatcher_, random_, local_info_, store_,
-                                  validation_visitor_, *api_),
-               EnvoyException);
-}
-
 TEST_F(CdsApiImplTest, Basic) {
-  interval_timer_ = new Event::MockTimer(&dispatcher_);
   InSequence s;
 
   setup();
@@ -375,20 +280,15 @@ resources:
     eds_config:
       path: eds path
 )EOF";
+  auto response1 = TestUtility::parseYaml<envoy::api::v2::DiscoveryResponse>(response1_yaml);
 
   EXPECT_CALL(cm_, clusters()).WillOnce(Return(ClusterManager::ClusterInfoMap{}));
   expectAdd("cluster1", "0");
   expectAdd("cluster2", "0");
   EXPECT_CALL(initialized_, ready());
-  EXPECT_CALL(*interval_timer_, enableTimer(_));
   EXPECT_EQ("", cds_->versionInfo());
-  EXPECT_EQ(0UL, cds_version_.value());
-  callbacks_->onSuccess(parseResponseMessageFromYaml(response1_yaml));
+  cds_callbacks_->onConfigUpdate(response1.resources(), response1.version_info());
   EXPECT_EQ("0", cds_->versionInfo());
-  EXPECT_EQ(7148434200721666028U, cds_version_.value());
-
-  expectRequest();
-  interval_timer_->callback_();
 
   const std::string response2_yaml = R"EOF(
 version_info: '1'
@@ -406,27 +306,24 @@ resources:
     eds_config:
       path: eds path
 )EOF";
+  auto response2 = TestUtility::parseYaml<envoy::api::v2::DiscoveryResponse>(response2_yaml);
 
   EXPECT_CALL(cm_, clusters()).WillOnce(Return(makeClusterMap({"cluster1", "cluster2"})));
   expectAdd("cluster1", "1");
   expectAdd("cluster3", "1");
   EXPECT_CALL(cm_, removeCluster("cluster2"));
-  EXPECT_CALL(*interval_timer_, enableTimer(_));
-  callbacks_->onSuccess(parseResponseMessageFromYaml(response2_yaml));
+  cds_callbacks_->onConfigUpdate(response2.resources(), response2.version_info());
 
-  EXPECT_EQ(2UL, store_.counter("cluster_manager.cds.update_attempt").value());
-  EXPECT_EQ(2UL, store_.counter("cluster_manager.cds.update_success").value());
   EXPECT_EQ("1", cds_->versionInfo());
-  EXPECT_EQ(13237225503670494420U, cds_version_.value());
 }
 
-TEST_F(CdsApiImplTest, Failure) {
-  interval_timer_ = new Event::MockTimer(&dispatcher_);
+// Validate behavior when the config is delivered but it fails PGV validation.
+TEST_F(CdsApiImplTest, FailureInvalidConfig) {
   InSequence s;
 
   setup();
 
-  const std::string response_yaml = R"EOF(
+  const std::string response1_yaml = R"EOF(
 version_info: '0'
 resources:
 - "@type": type.googleapis.com/envoy.api.v2.Cluster
@@ -442,25 +339,24 @@ resources:
     eds_config:
       path: eds path
 )EOF";
+  auto response1 = TestUtility::parseYaml<envoy::api::v2::DiscoveryResponse>(response1_yaml);
 
   EXPECT_CALL(cm_, clusters()).WillRepeatedly(Return(cluster_map_));
   EXPECT_CALL(initialized_, ready());
-  EXPECT_CALL(*interval_timer_, enableTimer(_));
-  callbacks_->onSuccess(parseResponseMessageFromYaml(response_yaml));
-
-  expectRequest();
-  interval_timer_->callback_();
-
-  EXPECT_CALL(*interval_timer_, enableTimer(_));
-
-  callbacks_->onFailure(Http::AsyncClient::FailureReason::Reset);
-
+  EXPECT_THROW(cds_callbacks_->onConfigUpdate(response1.resources(), response1.version_info()),
+               EnvoyException);
   EXPECT_EQ("", cds_->versionInfo());
-  EXPECT_EQ(2UL, store_.counter("cluster_manager.cds.update_attempt").value());
-  EXPECT_EQ(1UL, store_.counter("cluster_manager.cds.update_failure").value());
-  // Validate that the schema error increments update_rejected stat.
-  EXPECT_EQ(1UL, store_.counter("cluster_manager.cds.update_rejected").value());
-  EXPECT_EQ(0UL, cds_version_.value());
+}
+
+// Validate behavior when the config fails delivery at the subscription level.
+TEST_F(CdsApiImplTest, FailureSubscription) {
+  InSequence s;
+
+  setup();
+
+  EXPECT_CALL(initialized_, ready());
+  cds_callbacks_->onConfigUpdateFailed({});
+  EXPECT_EQ("", cds_->versionInfo());
 }
 
 } // namespace