+++ test/integration/integration_admin_test.cc
@@ -1,5 +1,7 @@
 #include "test/integration/integration_admin_test.h"
 
+#include <string>
+
 #include "envoy/admin/v2alpha/config_dump.pb.h"
 #include "envoy/config/metrics/v2/stats.pb.h"
 #include "envoy/http/header_map.h"
@@ -289,6 +291,13 @@ TEST_P(IntegrationAdminTest, Admin) {
   EXPECT_THAT(response->body(), testing::HasSubstr("added_via_api"));
   EXPECT_EQ("text/plain; charset=UTF-8", ContentType(response));
 
+  response = IntegrationUtil::makeSingleRequest(lookupPort("admin"), "GET", "/clusters?format=json",
+                                                "", downstreamProtocol(), version_);
+  EXPECT_TRUE(response->complete());
+  EXPECT_EQ("200", response->headers().Status()->value().getStringView());
+  EXPECT_EQ("application/json", ContentType(response));
+  EXPECT_NO_THROW(Json::Factory::loadFromString(response->body()));
+
   response = IntegrationUtil::makeSingleRequest(lookupPort("admin"), "POST", "/cpuprofiler", "",
                                                 downstreamProtocol(), version_);
   EXPECT_TRUE(response->complete());
@@ -342,17 +351,34 @@ TEST_P(IntegrationAdminTest, Admin) {
                                                 downstreamProtocol(), version_);
   EXPECT_TRUE(response->complete());
   EXPECT_EQ("200", response->headers().Status()->value().getStringView());
+  EXPECT_EQ("text/plain; charset=UTF-8", ContentType(response));
+  auto listeners = test_server_->server().listenerManager().listeners();
+  auto listener_it = listeners.cbegin();
+  for (; listener_it != listeners.end(); ++listener_it) {
+    EXPECT_THAT(response->body(), testing::HasSubstr(fmt::format(
+                                      "{}::{}", listener_it->get().name(),
+                                      listener_it->get().socket().localAddress()->asString())));
+  }
+
+  response = IntegrationUtil::makeSingleRequest(
+      lookupPort("admin"), "GET", "/listeners?format=json", "", downstreamProtocol(), version_);
+  EXPECT_TRUE(response->complete());
+  EXPECT_EQ("200", response->headers().Status()->value().getStringView());
   EXPECT_EQ("application/json", ContentType(response));
 
   json = Json::Factory::loadFromString(response->body());
-  std::vector<Json::ObjectSharedPtr> listener_info = json->asObjectArray();
+  std::vector<Json::ObjectSharedPtr> listener_info = json->getObjectArray("listener_statuses");
   auto listener_info_it = listener_info.cbegin();
-  auto listeners = test_server_->server().listenerManager().listeners();
-  auto listener_it = listeners.cbegin();
+  listeners = test_server_->server().listenerManager().listeners();
+  listener_it = listeners.cbegin();
   for (; listener_info_it != listener_info.end() && listener_it != listeners.end();
        ++listener_info_it, ++listener_it) {
-    EXPECT_EQ(listener_it->get().socket().localAddress()->asString(),
-              (*listener_info_it)->asString());
+    auto local_address = (*listener_info_it)->getObject("local_address");
+    auto socket_address = local_address->getObject("socket_address");
+    EXPECT_EQ(listener_it->get().socket().localAddress()->ip()->addressAsString(),
+              socket_address->getString("address"));
+    EXPECT_EQ(listener_it->get().socket().localAddress()->ip()->port(),
+              socket_address->getInteger("port_value"));
   }
 
   response = IntegrationUtil::makeSingleRequest(lookupPort("admin"), "GET", "/config_dump", "",